[BE/FEAT] 주간리포트, 결과화면 정확도>점수 수정

See merge request s12-ai-image-sub1/S12P21A103!118

+++ backend/src/main/java/com/prana/backend/yoga_sequence/controller/response/FeedbackResponse.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class FeedbackResponse {
     private String position;
-    private int feedbackCnt;
+    private Long feedbackCnt;
 }
